replaced remaining CreateUserFromRemoteIdentity with createFakeUser
This commit replaces the remaining uses of CreateUserFromRemoteIdentity
with the createFakeUser test helper function in web/*_test.go.
